I've started working on line 2472
#Comment when it's getting late. Say that the player should go to bed, say goodnight then close the game.
I've managed to get the basics to together. I'll try to make a second version with a feature that'll ask the player the time to set when it's getting late enough.
Most importantly of all, I need to get it together with the game's code. (here I didn't do anything other than adding a space at the end of the line 2472 (I'm pretty new to git hub and I wanted this announcement (if one can call it as such) as soon as possible since I'm already having some difficulty with finding words due to being very sleepy). I should continue the work after 9 hours once this is posted.