Update script-topics.rpy
#Comment when it's getting late. Say that the player should go to bed, say goodnight then close the game.
I've managed to get the basics to together. I'll try to make a second version with a feature that'll ask the player the time to set when it's getting late enough.
Most importantly of all, I need to get it together with the game's code. (I've added what I already have, however I'd like to be the only one working on this (the new code has hashtags in front so don't worry) since I've started it and I'm quite confident that I'll be able to finish without any major help. (I'm pretty new to git hub and I wanted this announcement (if one can call it as such) as soon as possible since I'm already having some difficulty with finding words due to being very sleepy). I should continue the work after 9 hours once this is posted.